refactor: separate concerns from controllers, ui adjustments
